post3

+++ app/Http/Controllers/TasksController.php
@@ -5,7 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 
 use App\Task;    // 追加
-use App\User; // 追加
+use App\User;    // 追加
 
 class TasksController extends Controller
 {
@@ -18,18 +18,19 @@ class TasksController extends Controller
     {
         $data = [];
         if (\Auth::check()) {
-            $tasks = Task::all();
-            
-            return view('tasks.index', [
+        $user = \Auth::user();
+        $tasks = $user->tasks()->orderBy('created_at', 'desc')->get();
+        
+        $data = [
+                'user' => $user,
                 'tasks' => $tasks,
-            ]);     //
+            ];
             
-        }else{
-        return view('welcome', $data); 
         }
+        return view('welcome', $data);
         
     }
-
+    
     /**
      * Show the form for creating a new resource.
      *
@@ -38,7 +39,7 @@ class TasksController extends Controller
     public function create()
     {
         $task = new Task;
-
+        
         return view('tasks.create', [
             'task' => $task,
         ]);
@@ -57,12 +58,13 @@ class TasksController extends Controller
             'content' => 'required|max:191',
         ]);
         
-        $task = new Task;
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
-        
+        $request->user()->tasks()->create([
+            'status' => $request->status,
+            'content' => $request->content,
+        ]);
+
         return redirect('/');
+
     }
 
     /**
@@ -73,11 +75,15 @@ class TasksController extends Controller
      */
     public function show($id)
     {
-        $task = Task::find($id);
-
-        return view('tasks.show', [
-            'task' => $task,
-        ]);
+         $task = \App\Task::find($id);
+         if (\Auth::id() === $task->user_id) {
+             
+             return view('tasks.show', [
+                 'task' => $task,
+            ]);
+         }
+        return redirect('/');
+        
         
     }
 
@@ -89,11 +95,15 @@ class TasksController extends Controller
      */
     public function edit($id)
     {
-        $task = Task::find($id);
-
-        return view('tasks.edit', [
-            'task' => $task,
-        ]);
+        $task = \App\Task::find($id);
+         if (\Auth::id() === $task->user_id) {
+             
+             return view('tasks.edit', [
+                 'task' => $task,
+            ]);
+         }
+        return redirect('/');
+        
     }
 
     /**
@@ -109,11 +119,14 @@ class TasksController extends Controller
             'status' => 'required|max:10',
             'content' => 'required|max:191',
         ]);
-        $task = Task::find($id);
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
 
+        $task = \App\Task::find($id);
+        
+        if (\Auth::id() === $task->user_id) {
+            $task->status = $request->status;
+            $task->content = $request->content;
+            $task->save();
+        }
         return redirect('/');
     }
 
@@ -125,9 +138,13 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        $task = Task::find($id);
-        $task->delete();
-
+        $task = \App\Task::find($id);
+        
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
+        
         return redirect('/');
     }
+    
 }
